break out block_metadata
